teach gozone about the $ORIGIN control entry
teach gozone about the $ORIGIN control entry, from RFC1035 s5.1
( https://tools.ietf.org/html/rfc1035 ).

This sets the default domain, which is filled in for domain names
specified as "@", and appended to relative domains (any domain not
ending in a ".").

This is the first control entry gozone has been taught about, which
means that it is now somewhat more stateful than it previously was.

Perhaps confusingly, we look for control entries when scanning the zone
file looking for Records, so scanning the "Next" record may also
influence the state in a way which will change the way subsequent
records are interpreted. This may be problematic, as one can also
specify the Origin via the new SetOrigin(string) method. Because control
entries are interpreted while scanning records, overriding an origin
specified before the first Record is not currently possible. The current
intended use-case for SetOrigin is to set the origin when no $ORIGIN
control entry exists at all.