Teach gozone about $TTL control entry
Previously, we allowed Records to not have a TTL specified, and we
stored the special value -1 to indicate that this was the case.

Now we teach gozone about the $TTL control entry, which specifies the
default for when this is the case.

I have not been able to find a reference where the format of the $TTL
control entry is specified, but I have seen it in the wild before, and
it is mentioned in the wikipedia page for Zone Files:

    https://en.wikipedia.org/wiki/Zone_file

At the moment, we only support the specification as a number, ie: not in
shorthand forms such as 1h = 3600. We may need to alter the Record
structure in order to support such specifications sanely (eg: storing
the TimeToLive as a string, and adding a new method for accessing the
TimeToLive). This may be warranted in any case, as now that we're adding
the handling of defaults, it becomes apparent that we may wish to make a
distinction between "specified" and "inherited" values.