allow backslash-escaping in non-"string" contexts
previously, we only considered data values to be "strings", and therefor
capable of backslash-escaping, when surrounded by double-quotes.
According to RFC1035, this escaping is actually possible in other
contexts as well, arguably all contexts.

Transitionally, we hard-code support for backslash-escaping into the
Default and Paren contexts. However, this points towards the need to
reconsider and potentially rewrite the main tokenizer, as our handling
of the Paren context, and the way Data is stored in general, is also
somewhat wanting.